requirements.txt
asgiref==3.2.7
Django==3.0.7
django-crispy-forms==1.9.1
django-simple-history==2.10.0
django-summernote==0.8.11.6
gunicorn==20.0.4
Pillow==6.2.1
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
